requirements.txt
fastapi
uvicorn
openai
ib_insync
psycopg2-binary
dotenv